requirements.txt
matplotlib==3.8.4
numpy==1.23.5
opencv-python==4.8.1.78
pandas==2.2.1
Pillow==9.5.0
plotly==5.20.0
pyrr==0.10.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.5.1
seaborn==0.13.2
sklearn==0.0
streamlit==1.41.1
sympy==1.12
tensorflow==2.12.1
tqdm==4.66.4
